+++ CHANGELOG.md
@@ -2,8 +2,52 @@
 
 ## [Unreleased](https://github.com/CosmWasm/wasmd/tree/HEAD)
 
-[Full Changelog](https://github.com/CosmWasm/wasmd/compare/v0.18.0...HEAD)
+[Full Changelog](https://github.com/CosmWasm/wasmd/compare/v0.20.0...HEAD)
 
+## [v0.20.0](https://github.com/CosmWasm/wasmd/tree/v0.20.0) (2021-10-08)
+
+[Full Changelog](https://github.com/CosmWasm/wasmd/compare/v0.20.0...v0.19.0)
+
+**Fixed bugs:**
+
+- Add capabilities to begin block [\#626](https://github.com/CosmWasm/wasmd/pull/626)
+
+**Api Breaking:**
+- Update to wasmvm 1.0.0-soon2 [\#624](https://github.com/CosmWasm/wasmd/issues/624)
+
+**Implemented Enhancements:**
+
+- Upgrade Cosmos-sdk v0.42.10 [\#627](https://github.com/CosmWasm/wasmd/pull/627) ([alpe](https://github.com/alpe))
+- Add transaction index implemented as counter [\#601](https://github.com/CosmWasm/wasmd/issues/601)
+- Fix inconsistent return of `contractAddress` from `keeper/init()`? [\#616](https://github.com/CosmWasm/wasmd/issues/616)
+- Query pinned wasm codes [\#596](https://github.com/CosmWasm/wasmd/issues/596)
+- Doc IBC Events [\#593](https://github.com/CosmWasm/wasmd/issues/593)
+- Allow contract Info query from the contract [\#584](https://github.com/CosmWasm/wasmd/issues/584)
+- Revisit reply gas costs for submessages. [\#450](https://github.com/CosmWasm/wasmd/issues/450)
+- Benchmarks for gas pricing [\#634](https://github.com/CosmWasm/wasmd/pull/634)
+- Treat all contracts as pinned for gas costs in reply [\#630](https://github.com/CosmWasm/wasmd/pull/630)
+- Bump github.com/spf13/viper from 1.8.1 to 1.9.0 [\#615](https://github.com/CosmWasm/wasmd/pull/615)
+
+## [v0.19.0](https://github.com/CosmWasm/wasmd/tree/v0.19.0) (2021-09-15)
+
+[Full Changelog](https://github.com/CosmWasm/wasmd/compare/v0.19.0...v0.18.0)
+
+**Fixed bugs:**
+
+- Ensure Queries are executed read only [\#610](https://github.com/CosmWasm/wasmd/issues/610)
+- Fix bug in query handler initialization on reply [\#604](https://github.com/CosmWasm/wasmd/issues/604)
+
+**Api Breaking:**
+- Bump Go version to  1.16 [\#612](https://github.com/CosmWasm/wasmd/pull/612)
+
+**Implemented Enhancements:**
+
+- Ensure query isolation [\#611](https://github.com/CosmWasm/wasmd/pull/611)
+- Optimize BalanceQuery [\#609](https://github.com/CosmWasm/wasmd/pull/609)
+- Bump wasmvm to v0.16.1 [\#605](https://github.com/CosmWasm/wasmd/pull/605)
+- Bump github.com/rs/zerolog from 1.23.0 to 1.25.0 [\#603](https://github.com/CosmWasm/wasmd/pull/603)
+- Add decorator options [\#598](https://github.com/CosmWasm/wasmd/pull/598)
+- Bump github.com/spf13/cast from 1.4.0 to 1.4.1 [\#592](https://github.com/CosmWasm/wasmd/pull/592)
 
 ## [v0.18.0](https://github.com/CosmWasm/wasmd/tree/v0.18.0) (2021-08-16)